Minor changes/refactor/version updates
