+++ web/package.json
@@ -3,24 +3,24 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "apollo-cache-inmemory": "^1.2.4",
+    "apollo-cache-inmemory": "^1.3.5",
     "apollo-cache-persist": "^0.1.1",
-    "apollo-client": "^2.3.2",
-    "apollo-link": "^1.2.2",
-    "apollo-link-error": "^1.1.0",
-    "apollo-link-http": "^1.5.4",
-    "apollo-link-queue": "^2.0.1",
-    "apollo-link-retry": "^2.2.4",
+    "apollo-client": "^2.4.2",
+    "apollo-link": "^1.2.3",
+    "apollo-link-error": "^1.1.1",
+    "apollo-link-http": "^1.5.5",
+    "apollo-link-queue": "^2.1.0",
+    "apollo-link-retry": "^2.2.5",
+    "react-apollo": "^2.2.4",
     "graphql": "^0.13.2",
     "graphql-tag": "^2.9.2",
     "moment": "^2.22.2",
     "pcmli.umbrella.core": "^8.0.1",
     "pcmli.umbrella.uni-core": "^0.7.18",
-    "react": "^16.4.1",
-    "react-apollo": "^2.1.4",
     "react-bootstrap": "^0.32.1",
-    "react-dom": "^16.4.1",
-    "react-scripts": "1.1.4",
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
+    "react-scripts": "2.0.5",
     "zen-observable": "^0.8.8"
   },
   "scripts": {
@@ -28,5 +28,11 @@
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
-  }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
