+++ web/src/index.js
@@ -3,9 +3,8 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 
-//import registerServiceWorker from './registerServiceWorker';
-// Do not cache anything
-/// registerServiceWorker();
+import registerServiceWorker from './registerServiceWorker';
+registerServiceWorker();
 
 ReactDOM.render(<App />, document.getElementById('root'));
